record and compare test output
